Cleanup: remove attachmentsPlugin.jsx and old editor code
